+++ src/LevelManager.cc
@@ -62,6 +62,10 @@ LevelManager::LevelManager(SimulationRunner *_runner, const bool _useLevels)
 
   this->ReadLevelPerformerInfo();
   this->CreatePerformers();
+
+  std::string service = "/world/";
+  service += this->runner->sdfWorld->Name() + "/level/add_performer";
+  this->node.Advertise(service, &LevelManager::OnAddPerformer, this);
 }
 
 /////////////////////////////////////////////////
@@ -166,9 +170,57 @@ void LevelManager::ReadPerformers(const sdf::ElementPtr &_sdf)
 
   if (this->useLevels && performerMap.empty())
   {
-    ignwarn << "Asked to use levels but no <performer>s were found - levels "
-               "will not work.\n";
+    ignwarn << "Asked to use levels but no <performer>s were found. Levels "
+               "will not work until performers are added.\n";
   }
+}
+
+/////////////////////////////////////////////////
+bool LevelManager::OnAddPerformer(const msgs::Performer &_req,
+                                  msgs::Boolean &_rep)
+{
+  _rep.set_data(false);
+  std::string name = _req.name();
+
+  // Find the model entity
+  Entity modelEntity = this->runner->entityCompMgr.EntityByComponents(
+      components::Name(name));
+  if (modelEntity == kNullEntity)
+  {
+    ignerr << "Unable to find model with name[" << name << "]. "
+      << "Performer not created\n";
+    return true;
+  }
+
+  // Check to see if the performer has already been added.
+  if (this->performerMap.find(name) == this->performerMap.end())
+  {
+    // Create the performer entity
+    Entity performerEntity = this->runner->entityCompMgr.CreateEntity();
+    this->performerMap[name] = performerEntity;
+
+    sdf::Geometry geometry(convert<sdf::Geometry>(_req.geometry()));
+    this->runner->entityCompMgr.CreateComponent(performerEntity,
+                                        components::Performer());
+    this->runner->entityCompMgr.CreateComponent(performerEntity,
+                                        components::PerformerLevels());
+    this->runner->entityCompMgr.CreateComponent(performerEntity,
+                                        components::Name("perf_" + name));
+    this->runner->entityCompMgr.CreateComponent(performerEntity,
+                                        components::Geometry(geometry));
+
+    // Make the model a parent of this performer
+    this->entityCreator->SetParent(this->performerMap[name], modelEntity);
+    _rep.set_data(true);
+  }
+  else
+  {
+    ignwarn << "Performer with name[" << name << "] "
+      << "has already been added.\n";
+  }
+
+  // The response succeeded.
+  return true;
 }
 
 /////////////////////////////////////////////////