+++ src/systems/state_publisher/StatePublisher.cc
@@ -15,8 +15,9 @@
  *
  */
 
-#include <ignition/plugin/Register.hh>
 #include <ignition/msgs/model.pb.h>
+
+#include <ignition/plugin/Register.hh>
 
 #include "ignition/gazebo/components/Name.hh"
 #include "ignition/gazebo/components/Joint.hh"